Don't install the package in edit mode to fix OptionConflictError
The entrypoints package (https://github.com/takluyver/entrypoints/) was introduced in flake8 3.7.

Unfortunately it will detect entrypoints multiple times when the plugin is installed in editable mode
because of this line: https://github.com/takluyver/entrypoints/blob/master/entrypoints.py#L192

The '-' is not in the directory name when you install a package in edit mode so when the python
path contains duplicate paths to the same module flake8 will instantiate the plugin multiple times
as well.