Increase required isort version to 4.3.5.
Having 'isort[pyproject] >= 4.3.0' as a requirement may make installation
fail, because isort versions 4.3.0 to 4.3.4 do not have the pyproject
extra, as it was added in version 4.3.5. Increasing the minimum version
to 4.3.5 should fix this.

Example of error when attempting to install flake8-isort:

pkg_resources.UnknownExtra: isort 4.3.4 has no such extra feature 'pyproject'