Let isort decide which config to use
isort will look for a configuration in pyproject.toml
when the TOML library is available.

flake8-isort has its own mechanism to look up an appropriate
configuration file and will pass its path on to isort.
Right now, flake8-isort doesn't consider pyproject.toml files as
configuration file candidates and will raise an error if there's
no other configuration file than the pyproject.toml.

Luckily, there's the parameter `--no-isort-config` to make flake8-isort
not stall when it cannot find a config file.
In this case, isort will not be instructed to look for a specific
configuration file by flake8-isort but instead look for it itself
(beginning from either the currently linted file path
or the current working directory).

This is great because it means that we can rely on isort
to figure out which config to use
and linting behavior between both libraries will be exactly the same.
The search for a configuration file can be removed from flake8-isort
and we get Pyproject.toml support for free.
The `--no-isort-config` option can be removed as well
because from now on it will always be the burden of isort
to figure out which config to use.

Related to #71
Closes #78