Use pip-tools to constrain CI requirements
In .travis.yml, the versions of coveralls and pytest-cov were
unconstrained, creating a risk of failure in Travis CI builds if an
incompatible version of pytest-cov or coveralls is released.

As suggested in #89, use pip-tools to address this risk. The
dependencies for coverage testing are in requirements-cov.in. The
frozen set of packages is in requirements-cov.txt, compiled for
python 2.7.

This is only for coverage testing, since pip-tools doesn't handle
relative paths and a minimal environment can help detect undeclared
dependencies.